requirements.txt
ultralytics==8.4.39
opencv-python==4.13.0.92
roboflow==1.3.3
torch==2.11.0
torchvision==0.26.0
numpy==2.2.6
pandas==2.3.3
Pillow==12.2.0
PyYAML==6.0.3
mediapipe==0.10.14
scikit-learn==1.7.2
joblib==1.5.3
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.3
python-multipart==0.0.26
websockets==16.0
SQLAlchemy==2.0.49
deep-translator==1.11.4
httpx==0.28.1
pytest==9.0.3
matplotlib==3.10.8
seaborn==0.13.2
